feat: created widget layer. created ToProfile Button, refactored paths & added aliases.
