+++ src/entities/viewer/ui/ToProfileButton/Ui.vue
@@ -1,0 +1,10 @@
+<template>
+  <q-btn
+    push
+    color="white"
+    text-color="primary"
+    round
+    icon="las la-user-circle"
+    :to="{ name: 'Profile' }"
+  />
+</template>
