+++ src/entities/viewer/ui/ToProfileButton/index.ts
@@ -1,0 +1,1 @@
+export { default as ToProfileButton } from './Ui.vue'
