--- src/entities/viewer/ui/profileForm/Ui.vue
@@ -1,77 +1,0 @@
-<template>
-  <q-page class="full-height column" padding>
-    <Form
-      :validation-schema="schema"
-      @submit="submit"
-    >
-      <Field
-        v-model="viewer.name"
-        name="name"
-        v-slot="{ errorMessage, value, field }"
-      >
-        <q-input
-          disable
-          :model-value="value"
-          v-bind="field"
-          :error-message="errorMessage"
-          :error="!!errorMessage"
-        />
-      </Field>
-
-      <Field
-        v-model="viewer.email"
-        name="email"
-        v-slot="{ errorMessage, value, field }"
-      >
-        <q-input
-          :model-value="value"
-          v-bind="field"
-          :error-message="errorMessage"
-          :error="!!errorMessage"
-        />
-      </Field>
-
-      <q-btn
-        class="q-mt-md"
-        color="primary"
-        type="submit"
-        label="Submit"
-      />
-    </Form>
-
-  </q-page>
-</template>
-
-<script lang="ts">
-import * as yup from 'yup';
-import { defineComponent } from 'vue';
-import { Field, Form } from 'vee-validate';
-import { viewerModel } from '@entities/viewer';
-import { Viewer } from '@shared/api';
-import { appNotify } from '@shared/notifications';
-
-export default defineComponent({
-  components: {
-    Field,
-    Form
-  },
-
-  setup() {
-    const viewer = viewerModel.store().viewer
-    const schema = yup.object({
-      username: yup.string().min(1),
-    });
-
-    const submit = async (dto:Partial<Viewer>) => {
-      await viewerModel.store().updateProfile(dto)
-      appNotify.success('Profile has been updated')
-    };
-
-    return {
-      viewer,
-      schema,
-      submit,
-    };
-  }
-});
-</script>
