+++ src/layouts/AuthLayout/ui/Index.vue
