+++ src/layouts/ViewerLayout/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as ViewerLayout } from './Index.vue'
