--- src/pages/Client/Profile/index.ts
@@ -1,1 +1,0 @@
-export {default as ClientProfile} from './Ui.vue'
