+++ src/pages/Client/index.ts
@@ -1,2 +1,1 @@
 export * from './Home'
-export * from './Profile'
