+++ src/router/routes.ts
@@ -1,38 +1,34 @@
 import { RouteRecordRaw } from 'vue-router';
-import { AuthLayout, ViewerLayout } from '@widgets/layouts';
+import { AuthLayout } from '@layouts/AuthLayout';
+import { ViewerLayout } from '@layouts/ViewerLayout';
 import { ClientHome } from '@pages/Client'
 import { CookHome } from '@pages/Cook'
 import { Login, Registration } from '@pages/Auth';
 import { Roles } from '@shared/api';
-import { ClientProfile } from '@pages/Client';
+import { Profile } from '@pages/Profile';
 
 const routes: RouteRecordRaw[] = [
   {
     path: '/',
     component: ViewerLayout,
-    meta: { requiresAuth: true, authorize: [Roles.client] },
+    meta: { requiresAuth: true, authorize: [Roles.client, Roles.cook] },
     children: [
       {
         path: '',
         name: 'ClientHome',
         component: ClientHome,
+        meta: { requiresAuth: true, authorize: [Roles.client] },
       },
-      {
-        path: '/profile',
-        name: 'Profile',
-        component: ClientProfile,
-      },
-    ],
-  },
-  {
-    path: '/',
-    component: ViewerLayout,
-    meta: { requiresAuth: true, authorize: [Roles.cook] },
-    children: [
       {
         path: '',
         name: 'CookHome',
         component: CookHome,
+        meta: { requiresAuth: true, authorize: [Roles.cook] },
+      },
+      {
+        path: 'profile',
+        name: 'Profile',
+        component: Profile,
       },
     ],
   },
