+++ src/shared/components/BakeryLogo.vue
@@ -1,0 +1,8 @@
+<template>
+  <div>
+    <q-avatar>
+      <img src="https://cdn.quasar.dev/logo-v2/svg/logo-mono-white.svg">
+    </q-avatar>
+    Quasar feature-sliced
+  </div>
+</template>
