+++ src/shared/components/index.ts
@@ -1,0 +1,1 @@
+export { default as BakeryLogo } from './BakeryLogo.vue'
