+++ src/widgets/TheHeader/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as TheHeader } from './Index.vue'
