--- src/widgets/layouts/ViewerLayout/ui/index.ts
@@ -1,2 +1,0 @@
-export { default as ViewerLayout } from './Index.vue'
-export { default as TheHeader } from './TheHeader.vue'
