--- src/widgets/layouts/index.ts
@@ -1,2 +1,0 @@
-export { AuthLayout } from './AuthLayout'
-export * from './ViewerLayout'
